requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
apify==2.5.0
apify_client==1.10.0
apify_fingerprint_datapoints==0.0.2
apify_shared==1.4.1
attrs==25.3.0
Brotli==1.1.0
browserforge==1.2.3
cachetools==5.5.2
camoufox==0.4.11
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
crawlee==0.6.9
cryptography==44.0.3
curl_cffi==0.11.1
eval_type_backport==0.2.2
filelock==3.18.0
frozenlist==1.6.0
geoip2==5.1.0
greenlet==3.2.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
language-tags==1.2.0
lazy-object-proxy==1.11.0
lxml==5.4.0
maxminddb==2.7.0
more-itertools==10.7.0
multidict==6.4.3
numpy==2.2.5
orjson==3.10.18
platformdirs==4.3.8
playwright==1.52.0
propcache==0.3.1
Protego==0.4.0
psutil==7.0.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.6.1
pydantic_core==2.33.2
pyee==13.0.0
PySocks==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
screeninfo==0.8.1
sniffio==1.3.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0
zstandard==0.23.0

fastapi~=0.115.12
starlette~=0.46.2
typing_extensions~=4.13.2
parsel~=1.10.0
pip~=25.1.1
wheel~=0.45.1
mysql-connector-python~=9.3.0
typing_extensions~=4.13.2